embedding_model/requirements.txt
toml
pydantic
pyarrow
pytorch
torchvision
scikit-learn
scipy
matplotlib
seaborn
pandas
polars
gensim